test(plugins): move plugin tests into separate files
